= Adjusted character type on windows
After some consideration, I've opted to change my recommended windows method back to _getch(). This is to ensure consistent performance, as I realized that the key mappings I do have (limited as they are) aren't even for wchar, and this can cause issues. A good example of this is the right-arrow key: With _getwch(), this can return -32 the first time and 77 each subsequent press. This is excellently specific... but not for my purposes here.